Change findMostSpecificMatch to return untouched match
ContentTypeUtils#findMostSpecificMatch returns a value as parsed by MediaType
instead of one of the values provided in the list of content types to match
against. However when MediaType parses a content type value with a parameter,
it is always transformed to a lowercase form.

This causes a problem when the response has a content type with an uppercase
parameter (e.g. "application/json;charset=UTF-8") because the value returned by
findMostSpecificMatch(e.g. "application/json;charset=utf-8") cannot be matched
to the API content while validating the response, resulting in a NPE.