Make schemas immutable by path
The Google JSON validator library caches schemas by their path. It assumes that
any reference it caches will remain valid forever - when the pointer references
a JsonNode that is later deleted, it throws a NullPointerException.

This manifests in NPEs bubbling up when performing multiple concurrent schema
validation operations.

This change makes DiscriminatorValidation subschemas immutable. They are
attached to the schema tree on the first visit to the discriminator validator
with a matching input instance. They remain there indefinitely. Thread IDs are
used for reentry checking to prevent infinite recursion.

This overall reduces the memory usage profile of the validator in the wild, as
before this change there would be one schema copy per concurrent validation,
and after this change there is one schema copy per distinct discriminator
value.